requirements.txt
torch
torch_geometric
rdkit
numpy
huggingface_hub[fsspec]
pyyaml
